+++ quantify_scheduler/backends/qblox/compiler_abc.py
@@ -25,6 +25,7 @@ from typing import (
     Union,
 )
 
+import numpy as np
 from pathvalidate import sanitize_filename
 from qcodes.utils.helpers import NumpyJSONEncoder
 from quantify_core.data.handling import gen_tuid, get_datadir
@@ -575,6 +576,39 @@ class Sequencer:
             acq.generate_data(wf_dict)
         return wf_dict
 
+    def _prepare_acq_settings(self, acquisitions: List[IOperationStrategy]):
+        """
+        Sets sequencer settings that are specific to certain acquisitions.
+        For example for a TTL acquisition strategy.
+
+        Parameters
+        ----------
+        acquisitions
+        """
+        acquisition_infos: List[OpInfo] = list(
+            map(lambda acq: acq.operation_info, acquisitions)
+        )
+        acq_metadata = _extract_acquisition_metadata_from_acquisitions(
+            acquisition_infos
+        )
+        if acq_metadata.acq_protocol == "trigger_count":
+            self._settings.ttl_acq_auto_bin_incr_en = (
+                acq_metadata.bin_mode == BinMode.AVERAGE
+            )
+            if self.connected_inputs is not None:
+                if len(self.connected_inputs) == 1:
+                    self._settings.ttl_acq_input_select = self.connected_inputs[0]
+                else:
+                    raise ValueError(
+                        f"Please make sure you use a single real input for this "
+                        f"portclock combination. "
+                        f"Found: {len(self.connected_inputs)} connected. "
+                        f"TTL acquisition does not support multiple inputs."
+                        f"Problem occurred for port {self.port} with"
+                        f"clock {self.clock}, which corresponds to {self.name} of "
+                        f"{self.parent.name}."
+                    )
+
     def _generate_acq_declaration_dict(
         self,
         acquisitions: List[IOperationStrategy],
@@ -636,7 +670,10 @@ class Sequencer:
             if acq_metadata.bin_mode == BinMode.APPEND:
                 num_bins = repetitions * (max(acq_indices) + 1)
             elif acq_metadata.bin_mode == BinMode.AVERAGE:
-                num_bins = max(acq_indices) + 1
+                if acq_metadata.acq_protocol == "trigger_count":
+                    num_bins = constants.MAX_NUMBER_OF_BINS
+                else:
+                    num_bins = max(acq_indices) + 1
             else:
                 # currently the BinMode enum only has average and append.
                 # this check exists to catch unexpected errors if we add more
@@ -920,16 +957,14 @@ class Sequencer:
         weights_dict = None
         acq_declaration_dict = None
         if self.parent.supports_acquisition:
-            weights_dict = (
-                self._generate_weights_dict() if len(self.acquisitions) > 0 else {}
-            )
-            acq_declaration_dict = (
-                self._generate_acq_declaration_dict(
+            weights_dict = {}
+            acq_declaration_dict = {}
+            if len(self.acquisitions) > 0:
+                self._prepare_acq_settings(self.acquisitions)
+                weights_dict = self._generate_weights_dict()
+                acq_declaration_dict = self._generate_acq_declaration_dict(
                     acquisitions=self.acquisitions, repetitions=repetitions
                 )
-                if len(self.acquisitions) > 0
-                else {}
-            )
 
         qasm_program = self.generate_qasm_program(
             self.parent.total_play_time,
@@ -1283,7 +1318,11 @@ class QbloxBaseModule(ControlDeviceCompiler, ABC):
             :code:`None` values.
         """
         underconstr = freqs.LO is None and freqs.IF is None
-        overconstr = False  # freqs.LO is not None and freqs.IF is not None, TODO FIXME
+        overconstr = (
+            freqs.LO is not None
+            and freqs.IF is not None
+            and not np.isclose(freqs.LO + freqs.IF, freqs.clock)
+        )
 
         if underconstr or overconstr:
             raise ValueError(