+++ CHANGELOG.md
@@ -19,8 +19,10 @@
 - Qblox Backend - Added method for gain configuration, overriding gain now raises ValueError (!533)
 - Qblox backend - add TriggerCount to `QRMAcquisitionManager` (!556)
 - Deprecation - Refactored tests to remove deprecated `qcompile`, refactored to `SerialCompiler` (!529, #368)
+- Qblox backend - Add preparation of acquisition settings and accompanying datastructures for NV centers (!567)
 - Waveforms - Fix `sudden_net_zero` waveform generation function misunderstands `amp_B` (!549, #390)
 - Qblox backend - Introduce `SetClockFrequency` operation, made possible by qblox-instruments==0.8 (!575, follow-up to !539 and !543)
+- Schedule Functions - make experiment-related schedule functions available in `quantify_scheduler.schedules` (!572)
 
 ## 0.10.1 (2022-12-20)